+++ src/app/admin/components/reports/reports.directive.spec.js
@@ -5,6 +5,23 @@
 
         var el, $log, $q, commonService, authService, vm, Mock;
 
+        var activityTemplate = {
+            acb: 'CCHIT',
+            date: 1492429771059,
+            certificationDate: 1285819200000,
+            certificationEdition: '2014',
+            csvDetails: undefined,
+            chplProductNumber: 'CHP-009351',
+            details: undefined,
+            developer: 'Epic Systems Corporation',
+            friendlyActivityDate: '2017-04-17',
+            friendlyCertificationDate: '2010-09-30',
+            id: 1480,
+            newId: 17497,
+            product: 'EpicCare Ambulatory - Core EMR 2',
+            questionable: false,
+        };
+
         beforeEach(function () {
             module('chpl.mock', 'chpl.templates', 'chpl.admin', function ($provide) {
                 $provide.decorator('authService', function ($delegate) {
@@ -108,6 +125,35 @@
             it('should not allow dates where start is after end', function () {
                 vm.activityRange.key.startDate = new Date('3/15/2017');
                 expect(vm.validDates('key')).toBe(false);
+            });
+        });
+
+        describe('ICS family activity', function () {
+            var icsFamilyActivity;
+            beforeEach(function () {
+                icsFamilyActivity = angular.copy(Mock.listingActivity[0]);
+            });
+
+            it('should recognize added/removed ICS Parents', function () {
+                icsFamilyActivity.originalData.ics.parents =[{chplProductNumber:'ID',certificationDate:1490194030517,certifiedProductId:1},{chplProductNumber:'ID2',certificationDate:1490194030517,certifiedProductId:2}];
+                icsFamilyActivity.newData.ics.parents = [{chplProductNumber:'ID2',certificationDate:1490194030517,certifiedProductId:2},{chplProductNumber:'ID3',certificationDate:1490194030517,certifiedProductId:3}];
+                var activity = angular.copy(activityTemplate);
+                activity.details = ['ICS Parent "ID" changes<ul><li>ID removed</li></ul>', 'ICS Parent "ID3" changes<ul><li>ID3 added</li></ul>'];
+                activity.csvDetails = 'ICS Parent "ID" changes<ul><li>ID removed</li></ul>\nICS Parent "ID3" changes<ul><li>ID3 added</li></ul>';
+
+                vm.interpretCps([icsFamilyActivity]);
+                expect(vm.searchedCertifiedProducts[0]).toEqual(activity);
+            });
+
+            it('should recognize added/removed ICS Children', function () {
+                icsFamilyActivity.originalData.ics.children = [{chplProductNumber:'ID2',certificationDate:1490194030517,certifiedProductId:2},{chplProductNumber:'ID3',certificationDate:1490194030517,certifiedProductId:3}];
+                icsFamilyActivity.newData.ics.children =[{chplProductNumber:'ID',certificationDate:1490194030517,certifiedProductId:1},{chplProductNumber:'ID2',certificationDate:1490194030517,certifiedProductId:2}];
+                var activity = angular.copy(activityTemplate);
+                activity.details = ['ICS Child "ID3" changes<ul><li>ID3 removed</li></ul>', 'ICS Child "ID" changes<ul><li>ID added</li></ul>'];
+                activity.csvDetails = 'ICS Child "ID3" changes<ul><li>ID3 removed</li></ul>\nICS Child "ID" changes<ul><li>ID added</li></ul>';
+
+                vm.interpretCps([icsFamilyActivity]);
+                expect(vm.searchedCertifiedProducts[0]).toEqual(activity);
             });
         });
     });